+++ web/members/testing.go
@@ -1,4 +1,4 @@
-package user
+package members
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/ssb-ngi-pointer/go-ssb-room/roomdb"
 )
 
-// MiddlewareForTests gives us a way to inject _test users_. It should not be used in production.
+// MiddlewareForTests gives us a way to inject _test members_. It should not be used in production.
 // This is part of testing.go because we need to use roomMemberContextKey, which shouldn't be exported either.
 // TODO: could be protected with an extra build tag.
 // (Sadly +build test does not exist https://github.com/golang/go/issues/21360 )
